remove README cruft from metadata.db
I'm not sure whether it's wrong, but it's definitely a bad idea to include monster amounts of readme info in metadata.db, *in-line*, and then refer back to it from Readme.md -- it gives us a cluttered, low-value metadata file and a useless Readme.  Let's just do the right thing, here.  Oh, and since the invention of whitespace in 4000bc it's been used for great gains in legibility, so that's been done a bit as well.